\name{checkFolderHierarchy}
\alias{checkFolderHierarchy}
\title{
	Check folder organization
}
\description{
	This function checks the folder organization of \code{indir} for the presence
	of folders used by WaveSeqR.
}
\usage{
checkFolderHierarchy(indir)
}
\arguments{
  \item{indir}{
	Input directory.
}
}
\author{
Apratim Mitra
}
\examples{

}
\keyword{utilities}